requirements.txt
Django==4.0.4
Pillow==9.1.1
asgiref==3.5.1
beautifulsoup4==4.11.1
cetifi==2022.5.18.1
charset-normalizer==2.0.12
django-bootstrap4==22.1
django-cors-headers==3.12.0
djangorestframework==3.13.1
idna==3.3
pip==22.0.4
pytz==2022.1
requests==2.27.1
setuptools==57.0.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
wheel==0.36.2